refactor: Unified classes
